fixed error when moving item cross device / fixed carousel content ordering

+++ crates/database/src/object.rs
@@ -6,6 +6,7 @@ use std::fs;
 use std::fs::File;
 use std::io::{BufReader, Read};
 use std::path::{Path, PathBuf};
+use std::process::Command;
 use tracing::{error};
 use types::database_ids::{ItemId, ObjectId};
 
@@ -41,12 +42,23 @@ impl Object {
         if !Object::data_path(new_object.id(), db).parent().unwrap().exists() {
             fs::create_dir_all(Object::data_path(new_object.id(), db).parent().unwrap())?;
         }
-        if let Err(err) = fs::rename(file, Object::data_path(new_object.id(), db)) {
+        if cfg!(unix) {
+            match Command::new("mv").arg(file).arg(Object::data_path(new_object.id(), db)).output() {
+                Ok(result) => {if !result.status.success() {
+                    return Err(Error::msg(format!("Failed to store new object : {}", std::str::from_utf8(result.stderr.as_slice())?)));
+                }}
+                Err(err) => {
+                    return Err(Error::msg(format!("Failed to store new object : {err}")));
+                }
+            }
+        } else {
+            if let Err(err) = fs::rename(file, Object::data_path(new_object.id(), db)) {
                 query_fmt!(db, r#"DELETE FROM SCHEMA_NAME.objects WHERE id = $1;"#, *new_object.id);
                 error!("Failed to rename object from {} to {}", file.display(), Object::data_path(new_object.id(), db).display());
                 let _ = fs::remove_file(file);
                 return Err(Error::msg(format!("Failed to store new object : {err} (please see server logs for more details) ")));
-        };
+            };
+        }
         Ok(new_object)
     }
